+++ collection_document_impl.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2017 ArangoDB GmbH, Cologne, Germany
+// Copyright 2017-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Ewout Prangsma
 //
 
 package driver
@@ -34,7 +32,7 @@ func (c *collection) DocumentExists(ctx context.Context, key string) (bool, erro
 	if err := validateKey(key); err != nil {
 		return false, WithStack(err)
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("HEAD", path.Join(c.relPath("document"), escapedKey))
 	if err != nil {
 		return false, WithStack(err)
@@ -55,7 +53,7 @@ func (c *collection) ReadDocument(ctx context.Context, key string, result interf
 	if err := validateKey(key); err != nil {
 		return DocumentMeta{}, WithStack(err)
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("GET", path.Join(c.relPath("document"), escapedKey))
 	if err != nil {
 		return DocumentMeta{}, WithStack(err)
@@ -240,7 +238,7 @@ func (c *collection) UpdateDocument(ctx context.Context, key string, update inte
 	if update == nil {
 		return DocumentMeta{}, WithStack(InvalidArgumentError{Message: "update nil"})
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("PATCH", path.Join(c.relPath("document"), escapedKey))
 	if err != nil {
 		return DocumentMeta{}, WithStack(err)
@@ -349,7 +347,7 @@ func (c *collection) ReplaceDocument(ctx context.Context, key string, document i
 	if document == nil {
 		return DocumentMeta{}, WithStack(InvalidArgumentError{Message: "document nil"})
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("PUT", path.Join(c.relPath("document"), escapedKey))
 	if err != nil {
 		return DocumentMeta{}, WithStack(err)
@@ -454,7 +452,7 @@ func (c *collection) RemoveDocument(ctx context.Context, key string) (DocumentMe
 	if err := validateKey(key); err != nil {
 		return DocumentMeta{}, WithStack(err)
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("DELETE", path.Join(c.relPath("document"), escapedKey))
 	if err != nil {
 		return DocumentMeta{}, WithStack(err)