+++ database_views_impl.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2018 ArangoDB GmbH, Cologne, Germany
+// Copyright 2018-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Ewout Prangsma
 //
 
 package driver
@@ -43,7 +41,7 @@ type getViewResponse struct {
 // View opens a connection to an existing view within the database.
 // If no collection with given name exists, an NotFoundError is returned.
 func (d *database) View(ctx context.Context, name string) (View, error) {
-	escapedName := pathEscape(name)
+	escapedName := pathEscape(name, d.conn)
 	req, err := d.conn.NewRequest("GET", path.Join(d.relPath(), "_api/view", escapedName))
 	if err != nil {
 		return nil, WithStack(err)
@@ -69,7 +67,7 @@ func (d *database) View(ctx context.Context, name string) (View, error) {
 
 // ViewExists returns true if a view with given name exists within the database.
 func (d *database) ViewExists(ctx context.Context, name string) (bool, error) {
-	escapedName := pathEscape(name)
+	escapedName := pathEscape(name, d.conn)
 	req, err := d.conn.NewRequest("GET", path.Join(d.relPath(), "_api/view", escapedName))
 	if err != nil {
 		return false, WithStack(err)