+++ edge_collection_documents_impl.go
@@ -52,7 +52,7 @@ func (c *edgeCollection) readDocument(ctx context.Context, key string, result in
 	if err := validateKey(key); err != nil {
 		return DocumentMeta{}, contextSettings{}, WithStack(err)
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("GET", path.Join(c.relPath(), escapedKey))
 	if err != nil {
 		return DocumentMeta{}, contextSettings{}, WithStack(err)
@@ -243,7 +243,7 @@ func (c *edgeCollection) updateDocument(ctx context.Context, key string, update 
 	if update == nil {
 		return DocumentMeta{}, contextSettings{}, WithStack(InvalidArgumentError{Message: "update nil"})
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("PATCH", path.Join(c.relPath(), escapedKey))
 	if err != nil {
 		return DocumentMeta{}, contextSettings{}, WithStack(err)
@@ -362,7 +362,7 @@ func (c *edgeCollection) replaceDocument(ctx context.Context, key string, docume
 	if document == nil {
 		return DocumentMeta{}, contextSettings{}, WithStack(InvalidArgumentError{Message: "document nil"})
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("PUT", path.Join(c.relPath(), escapedKey))
 	if err != nil {
 		return DocumentMeta{}, contextSettings{}, WithStack(err)
@@ -477,7 +477,7 @@ func (c *edgeCollection) removeDocument(ctx context.Context, key string) (Docume
 	if err := validateKey(key); err != nil {
 		return DocumentMeta{}, contextSettings{}, WithStack(err)
 	}
-	escapedKey := pathEscape(key)
+	escapedKey := pathEscape(key, c.conn)
 	req, err := c.conn.NewRequest("DELETE", path.Join(c.relPath(), escapedKey))
 	if err != nil {
 		return DocumentMeta{}, contextSettings{}, WithStack(err)