+++ id.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2017 ArangoDB GmbH, Cologne, Germany
+// Copyright 2017-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Ewout Prangsma
 //
 
 package driver
@@ -90,9 +88,4 @@ func (id DocumentID) Key() string {
 		return pathUnescape(parts[1])
 	}
 	return ""
-}
-
-// NewDocumentID creates a new document ID from the given collection, key pair.
-func NewDocumentID(collection, key string) DocumentID {
-	return DocumentID(pathEscape(collection) + "/" + pathEscape(key))
 }