+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Move examples to separate package
 - Deprecate ClientConfig.SynchronizeEndpointsInterval due to bug in implementation
 - [V1] Add Rename function for collections (single server only).
+- [V1] Do not escape strings params in URL in case of VST connection
 
 ## [1.6.0](https://github.com/arangodb/go-driver/tree/v1.6.0) (2023-05-30)
 - Add ErrArangoDatabaseNotFound and IsExternalStorageError helper to v2